Revert "Revert "essaie de correction de la suppression des posts""
This reverts commit f40999dd54b04d265873907cb680048f968737ae.

+++ API/controllers/post.js
@@ -2,12 +2,16 @@ const Post = require('../models/post');
 const fs = require('fs');
 
 exports.createPost = (req, res, next) => {
-    const postObject = JSON.parse(req.body.post);
+    const postObject = req.body.post;
     delete postObject._id;
     const post = new Post({
         ...postObject,
-        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+        createdDate: Date.now(),
+        user: req.auth.userId,
     });
+    if (req.file) {
+        post.imageUrl = `${req.protocol}://${req.get('host')}/images/${req.file.filename}`;
+    }
     post.save()
         .then(() => res.status(201).json({message: 'Nouveau post créé'}))
         .catch(error => res.status(401).json({error}));
@@ -15,7 +19,8 @@ exports.createPost = (req, res, next) => {
 
 exports.getAllPosts = (req, res, next) => {
     Post.find()
-        .then(posts = res.status(200).json(posts))
+        .populate('user')
+        .then(posts => res.status(200).json(posts))
         .catch(error => res.status(400).json({error}))
 };
 
@@ -49,23 +54,33 @@ exports.likePost = (req, res, next) => {
 exports.deletePost = (req, res, next) => {
     Post.findOne({_id: req.params.id})
         .then(post => {
-            const filename = post.imageUrl.split('/images')[1];
-            fs.unlink(`images/${filename}`, () => {
-                Post.deleteOne({_id: req.params})
-                    .then(() => res.status(200).json({message: 'Post supprimé'}))
-                    .catch(error => res.status(400).json({error}));
-            });
+            if (post.user.valueOf() != req.auth.userId && !req.auth.admin) {
+                res.status(401).json({message: 'Not authorized'});
+            }
+            else {
+                const filename = post.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Post.deleteOne({_id: req.params.id})
+                        .then(() => res.status(200).json({message: 'Post supprimé'}))
+                        .catch(error => res.status(400).json({error}));
+                });
+            }
         })
-        .catch(error => res.status(500).json({error}));
+        //.catch(error => res.status(500).json({error}));
 };
 
 exports.modifyPost = (req, res, next) => {
-    const postObject = req.file ?
-        {
-            ...JSON.parse(req.body.post),
-            imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
-        } : {...req.body};
-    Post.updateOne({_id: req.params.id}, {...postObject, _id: req.params.id})
-        .then(() => res.status(200).json({message: 'Post mis à jour !'}))
-        .catch(error => res.status(400).json({error}));
+    if (post.userId != req.params.id && !req.auth.admin) {
+        res.status(401).json({message: 'Not authorized'});
+    }
+    else {
+        const postObject = req.file ?
+            {
+                ...JSON.parse(req.body.post),
+                imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+            } : {...req.body};
+        Post.updateOne({_id: req.params.id}, {...postObject, _id: req.params.id})
+            .then(() => res.status(200).json({message: 'Post mis à jour !'}))
+            .catch(error => res.status(400).json({error}));
+    }
 };
